update scripts
update mac install script brew.sh; organize and refactor code

+++ scripts/brew.sh
@@ -1,20 +1,35 @@
 #!/bin/bash
 
 is_app_installed() {
-	type "$1" &>/dev/null
+	command -v "$1" &>/dev/null
 }
 
-# If Homebrew is not installed, install Homebrew
-if test ! "$( command -v brew )"; then
+log_table() {
+	items=("$@")
+	for item in "${items[@]}"; do
+		printf "%-8s\n" "${item}"
+	done | column
+}
+
+echo "Checking if xcode is installed..."
+if ! is_app_installed xcode-select; then
+	echo "xcode-select is not installed, installing xcode-select..."
+	# Install the command line developer toolsc
+	xcode-select --install
+else
+	echo "xcode-select is installed!"
+fi
+
+echo "Checking if homebrew is installed..."
+if ! is_app_installed brew; then
 	echo "Homebrew not installed. Installing homebrew..."
-	ruby -e "$( crl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install )"
+	/bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
+else
+	echo "Homebrew is installed!"
 fi
 
 echo -e "\\n\\nInstalling homebrew packages..."
 echo "================================"
-
-# Install the command line developer toolsc
-xcode-select --install
 
 install_packages() {
 	# Specify brew packages to be installed
@@ -22,7 +37,6 @@ install_packages() {
 		archey #screenfetch
 		cask #macOS GUI app installer
 		cmake
-		ctags
 		deno
 		fzf
 		gcc #need to compile cask
@@ -30,15 +44,14 @@ install_packages() {
 		htop #resource monitor
 		irssi #irc
 		links #terminal browser
-		macvim
 		mas
 		ncdu #disk util
 		neovim
 		nmap #network scanner
-		node
 		nvm
 		peco #interactive filtering
 		python
+		reattach-to-user-namespace
 		ripgrep
 		tig #git
 		tmux
@@ -51,69 +64,40 @@ install_packages() {
 	guiApps=(
 		docker
 		visual-studio-code
-		macvim
 		google-chrome
 		dropbox
 	)
 
-	packages_length=${#cmlApps[@]}
-	current_pos=1
-	packages_labels=""
+	echo "Command line apps: "
+	log_table "${cmlApps[@]}"
 
-	for package in "${cmlApps[@]}"
-	do
-		if [[ "$current_pos" -eq "$packages_length" ]]; then
-			packages_labels+="${package}"
-		else
-			packages_labels+="${package}, "
-		fi
+	read -p "Do you want to install essential workflow packages listed above? (y/n) " should_install_packages
 
-		current_pos=$((current_pos + 1))
-	done
-
-	echo "Command line apps: "${packages_labels}
-	read -p "Do you want to install essential workflow packages? (y/n) " should_install_packages
-
-	if [ $should_install_packages = 'y' ]
-	then
+	if [ $should_install_packages = 'y' ]; then
 		# Install specified packages, skipping installed packages
 		for app in "${cmlApps[@]}"; do
 			appName=$( echo "$app" | awk '{print $1}' )
 			if brew list "$appName" > /dev/null 2>&1; then
 				echo "$appName already installed... skipping."
 			else
-				brew install "$app"
+				HOMEBREW_NO_AUTO_UPDATE=1 brew install "$app"
 			fi
 		done
 	fi
 
-	packages_length=${#guiApps[@]}
-	current_pos=1
-	packages_labels=""
+	echo "GUI apps: "
+	log_table "${guiApps[@]}"
 
-	for package in "${guiApps[@]}"
-	do
-		if [[ "$current_pos" -eq "$packages_length" ]]; then
-			packages_labels+="${package}"
-		else
-			packages_labels+="${package}, "
-		fi
+	read -p "Do you want to install essential workflow gui apps listed above? (y/n) " should_install_gui_apps
 
-		current_pos=$((current_pos + 1))
-	done
-
-	echo "GUI apps: "${packages_labels}
-	read -p "Do you want to install essential workflow packages? (y/n) " should_install_gui_apps
-
-	if [ $should_install_gui_apps = 'y' ]
-	then
+	if [ $should_install_gui_apps = 'y' ]; then
 		# Install specified packages, skipping installed packages
 		for app in "${guiApps[@]}"; do
 			appName=$( echo "$app" | awk '{print $1}' )
 			if brew cask list "$appName" > /dev/null 2>&1; then
 				echo "$appName already installed... skipping."
 			else
-				brew install --cask "$app"
+				HOMEBREW_NO_AUTO_UPDATE=1 brew install --cask "$app"
 			fi
 		done
 	fi
@@ -122,8 +106,7 @@ install_packages() {
 install_oh_my_zsh() {
 	read -p "Do you want to install oh-my-zsh? (y/n) " should_install_oh_my_zsh
 
-	if [ $should_install_oh_my_zsh = 'y' ]
-	then
+	if [ $should_install_oh_my_zsh = 'y' ]; then
 		echo "Installing oh-my-zsh..."
 		sh -c "$(curl -fsSL https://raw.githubusercontent.com/robbyrussell/oh-my-zsh/master/tools/install.sh)"
 	fi
@@ -132,8 +115,7 @@ install_oh_my_zsh() {
 make_zsh_default_shell() {
 	read -p "Do you want to make zsh the default shell? (y/n) " should_make_zsh_default_shell
 
-	if [ $should_make_zsh_default_shell = 'y' ]
-	then
+	if [ $should_make_zsh_default_shell = 'y' ]; then
 		chsh -s $(which zsh)
 	fi
 }
@@ -141,8 +123,7 @@ make_zsh_default_shell() {
 install_vimplug_vim() {
 	read -p "Do you want to install vimplug for vim? (y/n) " should_install_vimplug_vim
 
-	if [ $should_install_vimplug_vim = 'y' ]
-	then
+	if [ $should_install_vimplug_vim = 'y' ]; then
 		echo "Installing vimplug (vim)..."
 		curl -fLo ~/.vim/autoload/plug.vim --create-dirs https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
 	fi
@@ -151,8 +132,7 @@ install_vimplug_vim() {
 install_vimplug_neovim() {
 	read -p "Do you want to install vimplug for neovim? (y/n) " should_install_vimplug_neovim
 
-	if [ $should_install_vimplug_neovim = 'y' ]
-	then
+	if [ $should_install_vimplug_neovim = 'y' ]; then
 		echo "Installing vimplug (neovim)..."
 		curl -fLo ~/.local/share/nvim/site/autoload/plug.vim --create-dirs https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
 	fi
@@ -161,8 +141,7 @@ install_vimplug_neovim() {
 install_tmux_plugin_manager() {
 	read -p "Do you want to install tmux_plugin_manager? (y/n) " should_install_tmux_plugin_manager
 
-	if [ $should_install_tmux_plugin_manager = 'y' ]
-	then
+	if [ $should_install_tmux_plugin_manager = 'y' ]; then
 		echo "Installing tmux plugin manager"
 		git clone https://github.com/tmux-plugins/tpm ~/.config/tmux/plugins/tpm
 	fi
@@ -187,15 +166,29 @@ if is_app_installed tmux; then
 	install_tmux_plugin_manager
 fi
 
-# After the install, setup fzf
-echo -e "\\n\\nRunning fzf install script..."
-echo "================================"
-/usr/local/opt/fzf/install --all
+read -p "Do you want to run fzf install script? (y/n) " should_run_fzf_install_script
 
-# After the install, install neovim python libraries
-echo -e "\\n\\nRunning Neovim Python install..."
-echo "================================"
-pip3 install pynvim
+if [ $should_run_fzf_install_script = 'y' ]; then
+	if [ -f /usr/local/opt/fzf/install ]; then
+		if ! is_app_installed fzf; then
+			# After the install, setup fzf
+			echo -e "\\n\\nRunning fzf install script..."
+			echo "================================"
+			/usr/local/opt/fzf/install --all
+		fi
+	else
+		echo "fzf install script not found. Did not install fzf."
+	fi
+fi
+
+read -p "Do you want to install pynvim? (y/n) " should_install_pynvim
+
+if [ $should_install_pynvim = 'y' ]; then
+	# After the install, install neovim python libraries
+	echo -e "\\n\\nRunning Neovim Python install..."
+	echo "================================"
+	pip3 install pynvim
+fi
 
 echo "brew.sh script finished."
